backend/requirements.txt
flask
flask-cors
gunicorn
pandas
numpy
faiss-cpu
sentence-transformers
python-dotenv
bcrypt
requests
psycopg2-binary
boto3


